feat: 재판예약 완료 페이지 템플릿 완성
