+++ src/interfaces/ButtonItem.ts
@@ -1,4 +1,5 @@
 export interface ButtonItem {
   text: string;
   size: string;
+  onClick: any;
 }
